Rename '0-global.js' to 'common.js'
